backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
experta
pyyaml
python-multipart
httpx
pytest